+++ crates/swc_ecma_minifier/src/compress/pure/conds.rs
@@ -2,10 +2,14 @@ use std::mem::swap;
 
 use swc_common::{util::take::Take, EqIgnoreSpan};
 use swc_ecma_ast::*;
-use swc_ecma_utils::{ExprExt, Type, Value};
+use swc_ecma_utils::{ExprExt, StmtExt, StmtLike, Type, Value};
 
 use super::Pure;
-use crate::{compress::util::negate_cost, debug::dump, util::make_bool};
+use crate::{
+    compress::util::{negate, negate_cost},
+    debug::dump,
+    util::make_bool,
+};
 
 impl Pure<'_> {
     ///
@@ -177,5 +181,76 @@ impl Pure<'_> {
                 }
             }
         }
+    }
+
+    /// if (foo) return bar()
+    /// else baz()
+    ///
+    /// `else` token can be removed from the code above.
+    pub(super) fn drop_else_token<T>(&mut self, stmts: &mut Vec<T>)
+    where
+        T: StmtLike,
+    {
+        // Find an if statement with else token.
+        let need_work = stmts.iter().any(|stmt| match stmt.as_stmt() {
+            Some(Stmt::If(IfStmt {
+                cons,
+                alt: Some(..),
+                ..
+            })) => cons.terminates(),
+            _ => false,
+        });
+        if !need_work {
+            return;
+        }
+        // TODO(kdy1): Move this to pure optimizer and handle termination within a pass.
+        // let mut need_terminate_check = false;
+        //
+
+        let mut new_stmts = vec![];
+
+        for stmt in stmts.take() {
+            match stmt.try_into_stmt() {
+                Ok(stmt) => match stmt {
+                    Stmt::If(IfStmt {
+                        span,
+                        mut test,
+                        mut cons,
+                        alt: Some(mut alt),
+                        ..
+                    }) if cons.terminates() => {
+                        if let (
+                            Stmt::Return(ReturnStmt { arg: None, .. }),
+                            Stmt::Decl(Decl::Fn(..)),
+                        ) = (&*cons, &*alt)
+                        {
+                            // I don't know why, but terser behaves differently
+                            negate(&self.expr_ctx, &mut test, true, false);
+
+                            swap(&mut cons, &mut alt);
+                            // need_terminate_check = true;
+                        }
+
+                        new_stmts.push(T::from_stmt(Stmt::If(IfStmt {
+                            span,
+                            test,
+                            cons,
+                            alt: None,
+                        })));
+                        new_stmts.push(T::from_stmt(*alt));
+                    }
+                    _ => {
+                        new_stmts.push(T::from_stmt(stmt));
+                    }
+                },
+                Err(stmt) => new_stmts.push(stmt),
+            }
+        }
+
+        // if need_terminate_check {}
+
+        self.changed = true;
+        report_change!("conditionals: Dropped useless `else` token");
+        *stmts = new_stmts;
     }
 }
